Version 0.27.0 - Minor corrections in Auth and Profile tests - Minor adjustments in Slug tests - Added portfolio tests: to verify new project creation and to manually test database failure during project saving - Added a condition in AppServiceProvider to check if the pages table exists before caching pages during application bootstrap - Wrapped the images()->create() call inside the imageUpload method of the Portfolio model in a try/catch block and re-threw the exception to allow proper handling during database transaction failures - Minor correction in the conditional slug rule in StorePortfolioRequest

+++ app/Http/Controllers/Admin/PortfolioController.php
@@ -9,7 +9,9 @@ use App\Services\CacheCleanerService;
 use App\Traits\HasThumbnail;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Storage;
+use mysql_xdevapi\Exception;
 
 
 class PortfolioController extends Controller
@@ -52,28 +54,30 @@ class PortfolioController extends Controller
         $savedImagePaths = [];
 
         try {
-
             DB::transaction(function () use ($validated, $images, $positions, &$savedImagePaths) {
 
                 $portfolio = Portfolio::create($validated);
 
                 $savedImagePaths = $portfolio->saveImages($positions, $images);
-
                 $portfolio->slugs()->create([
                     'slug' => $validated['slug'],
                     'is_current' => true
                 ]);
-
             });
 
         } catch (\Throwable $e) {
-            Storage::delete($validated['thumbnail']);
+
+            if($validated['thumbnail']){
+                Storage::delete($validated['thumbnail']);
+            }
 
             foreach ($savedImagePaths as $path) {
                 Storage::delete($path);
             }
 
-            throw $e;
+            Log::error('Portfolio store failed: ' . $e->getMessage());
+            return back()->with('error', 'Portfolio store failed.');
+
         }
 
         CacheCleanerService::cleanCache([
@@ -92,7 +96,6 @@ class PortfolioController extends Controller
     public function showSingle(Portfolio $portfolio)
     {
         return view('admin.portfolio.preview-single', ['portfolio' => $portfolio]);
-//        TODO: Change to slug in URL
     }
 
     public function showGrid(Portfolio $portfolio){
@@ -178,14 +181,14 @@ class PortfolioController extends Controller
 
             });
         }catch (\Throwable $e){
-            Storage::delete($validated['thumbnail']);
 
-            // remove images
+            // remove new images
             foreach ($updatedImagePaths as $path) {
                 Storage::delete($path);
             }
 
-            throw $e;
+            Log::error('Portfolio update failed: ' . $e->getMessage());
+            return back()->with('error', 'Portfolio store failed.');
         }
 
         $portfolio->deleteImages($existingImages, $oldImagesIds);
